requirements.txt
django==4.1.7
django-environ==0.9.0
Pillow==9.4.0
psycopg2-binary==2.9.5
djangorestframework==3.14.0
django-bootstrap-v5==1.0.11
django-bootstrap-icons==0.8.2